refactor: restructure hover components and remove unused files
- Deleted LrcTab, TimeTab, and WeatherTab components as part of a cleanup.
- Moved HoverContent and HoverForm components to a new structure for better organization.
- Updated index export paths to reflect the new component locations.
- Recreated LrcTab, TimeTab, and WeatherTab components in the new structure.
