+++ src/renderer/components/states/hover/pages/LrcTab.tsx
