+++ src/renderer/components/states/hover/pages/WeatherTab.tsx
